finished writing task usecase test and succedded

+++ test-go-backend-task-manager/usecases/task_usecases.go
@@ -37,6 +37,9 @@ func (taskUC taskUseCase) GetTaskByID(cxt context.Context, taskID string) (domai
 func (taskUC *taskUseCase) CreateTask(cxt context.Context, newTask domain.Task) (string, *domain.TaskError) {
 	context, cancel := context.WithTimeout(cxt, taskUC.contextTimeout)
 	defer cancel()
+	if newTask.Title == "" {
+		return "", &domain.TaskError{Message: "Title is required", Code: 400}
+	}
 
 	return taskUC.taskRepository.CreateTask(context, newTask)
 }
@@ -48,9 +51,18 @@ func (taskUC *taskUseCase) UpdateTask(cxt context.Context, updateTask domain.Tas
 	return taskUC.taskRepository.UpdateTask(context, updateTask)
 }
 
-func (taskUC *taskUseCase) DeleteTask(cxt context.Context, taskID string) (domain.Task, *domain.TaskError) {
+func (taskUC *taskUseCase) DeleteTask(cxt context.Context, taskID string, authority domain.User) (domain.Task, *domain.TaskError) {
 	context, cancel := context.WithTimeout(cxt, taskUC.contextTimeout)
 	defer cancel()
+
+	fetchedTask, errFetch := taskUC.taskRepository.FetchTaskByID(context, taskID)
+	if errFetch != nil {
+		return domain.Task{}, errFetch
+	}
+
+	if authority.ID != fetchedTask.UserID {
+		return domain.Task{}, &domain.TaskError{Message: "You are not authorized to delete this task", Code: 403}
+	}
 
 	return taskUC.taskRepository.DeleteTask(context, taskID)
 }
